requirements.txt
torch
torchvision
torchaudio
ftfy
regex
tqdm
easydict
PyYAML
nltk
numpy
matplotlib
transformers
pandas
scikit-learn
scipy
wandb
Pillow
gdown
thinc==8.2.5
spacy==3.7.6
git+https://github.com/openai/CLIP.git
Cython
git+https://github.com/waleedka/coco.git#subdirectory=PythonAPI